Create index.html
This is a simple form to collect information on the users, name, company, email, and postcode along with a little message, displayed across the screen.

The placeholder attribute is so that labels appear inside the input fields quite nice to minimising screen real estate. The HTML5 attribute required is used, however HTML5 validation occurs upon submit which is not performed so the jQuery validation kicks in first and ajax takes care of submitting data to the php file rather than the form submit.

An empty panel exists and starts off hidden for the purposes of displaying the return message from the PHP file indicating success/failure of form data emailed.